% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootCoefQuartVar.R
\docType{data}
\name{BootCoefQuartVar}
\alias{BootCoefQuartVar}
\title{R6 Bootstrap Resampling for Coefficient of Quartile Variation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
\code{BootCoefQuartVar$new(x, ...)}

## Default R6 method:
\code{BootCoefQuartVar$new(x, na.rm = FALSE, digits = 4,
               R = 1000, ...)$boot_cqv()}
}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for numeric vectors}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{R}{integer indicating the number of bootstrap replicates.}
}
\description{
The R6 class \code{BootCoefQuartVar} produces the bootstrap
             resamplimg for the coeficient of quartile variation (cqv) of the
             given numeric vectors. It uses \link[boot]{boot} from the
             package \pkg{boot}.
}
\examples{
x <- c(
    0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
    4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
cqv_x <- BootCoefQuartVar$new(x, na.rm = TRUE, digits = 3, R = 1000)
cqv_x$boot_cqv()
R6::is.R6(cqv_x)
}
\keyword{datasets}
