% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqv.R
\name{cqv}
\alias{cqv}
\title{Coefficient of Quartile Variation (CQV)}
\usage{
cqv(x, na.rm, digits)
}
\arguments{
\item{x}{An {R} object. Currently there are methods for numeric vectors}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\description{
Generic function for the coefficient of quartile variation (cqv)
}
\examples{
x <- rnorm(100, mean = 10, sd = 6)
cqv(x, na.rm = TRUE, digits = 2)
}
\references{
Bonett, DG., 2006, Confidence interval for a coefficient of quartile variation, Computational Statistics & Data Analysis, 50(11), 2953-7, DOI: \href{https://doi.org/10.1016/j.csda.2005.05.007}{https://doi.org/10.1016/j.csda.2005.05.007}
}
