% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantIndex.R
\docType{data}
\name{QuantIndex}
\alias{QuantIndex}
\title{R6 Sample Quantiles}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
\code{QuantIndex$new(x, ...)}

## Default R6 method:
\code{QuantIndex$new(x, na.rm = TRUE, digits = 4,
               probs = 0.5, ...)$qx()}
}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for numeric vectors}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{probs}{numeric vector of probabilities with values in \code{[0,1]}.}
}
\description{
The R6 class \code{QuantIndex} produces the sample quantiles
             corresponding to the given probabilities. It uses
             \link[stats]{quantile} from the package \pkg{stats}.
}
\examples{
x <- c(
    0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
    4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
QuantIndex$new(x, na.rm = TRUE, digits = 4, probs = 0.5)$qx()
quant95 <-  QuantIndex$new(x, na.rm = TRUE, digits = 4, probs = 0.95)
quant95$qx()
R6::is.R6(quant95)
}
\keyword{datasets}
