% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.versatile.R
\name{rm.versatile}
\alias{rm.versatile}
\title{Versatile Remove Objects}
\usage{
rm.versatile(save.objects = NULL, save.patterns = NULL,
  rm.objects = NULL, rm.patterns = NULL, modes = list("integer",
  "double", "character", "list"), envir = .GlobalEnv, inherits = FALSE)
}
\arguments{
\item{save.objects}{an optional list naming objects to be saved (\emph{i.e.,}
not to be removed).}

\item{save.patterns}{x}
}
\description{
Versatile function to remove objects.
}
