% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqv.R
\name{cqv}
\alias{cqv}
\title{Coefficient of Quartile Variation (CQV)}
\usage{
cqv(x, na.rm = FALSE, digits = NULL, CI = NULL, R = NULL, ...)
}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for numeric vectors}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{CI}{a scalar representing the type of confidence intervals required.
The value should be any of the values "Bonett", "norm","basic", "
perc", "bca" or "all".}

\item{R}{integer indicating the number of bootstrap replicates.}
}
\value{
An object of type "list" which contains the cqv, the
        intervals, and the computation method. It has these components:

\describe{
       \item{$method}{
       A description of statistical method used for the computations.
       }
       \item{$statistics}{
       A data frame representing three vectors: cqv, lower and upper limits
       of 95\% confidence interval \code{(CI)}.
       }
       }
}
\description{
Generic function for the coefficient of quartile variation (cqv)
}
\details{
\describe{
        \item{\strong{Coefficient of Quartile Variation}}{
        \code{\deqn{ cqv = ((q3-q1)/(q3 + q1))*100 , } } where \eqn{q3}
        and \eqn{q1} are third quartile (\emph{i.e.,} 75th percentile) and
        first quartile (\emph{i.e.,} 25th percentile), respectively.
        The \emph{cqv} is a measure of relative dispersion that is based on
        interquartile range \emph{(iqr)}. Since \eqn{cqv} is unitless, it
        is useful for comparison of variables with different units. It is
        also a measure of homogeneity [1, 2].
        }
        }
}
\examples{
x <- c(
    0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
    4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
cqv(x)
cqv(x, na.rm = TRUE, digits = 2)
cqv(x, na.rm = TRUE, digits = 2, CI = "Bonett")
cqv(x, na.rm = TRUE, digits = 2, CI = "norm")
cqv(x, na.rm = TRUE, digits = 2, CI = "basic")
cqv(x, na.rm = TRUE, digits = 2, CI = "perc")
cqv(x, na.rm = TRUE, digits = 2, CI = "bca")
cqv(x, na.rm = TRUE, digits = 2, CI = "all")

}
\references{
[1] Bonett, DG., 2006, Confidence interval for a coefficient of quartile variation, Computational Statistics & Data Analysis, 50(11), 2953-7, DOI: \href{https://doi.org/10.1016/j.csda.2005.05.007}{https://doi.org/10.1016/j.csda.2005.05.007}

[2] Altunkaynak, B., Gamgam, H., 2018, Bootstrap confidence intervals for the coefficient of quartile variation, Simulation and Computation, 1-9, DOI: \href{https://doi.org/10.1080/03610918.2018.1435800}{https://doi.org/10.1080/03610918.2018.1435800}
}
