% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation (cv)}
\usage{
cv(x, na.rm = FALSE, digits = NULL, method = NULL, unbiased = TRUE,
  alpha = 0.05, R = NULL, ...)
}
\description{
Generic function for the coefficient of variation (cv)
}
