% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.versatile.R
\name{rm.versatile}
\alias{rm.versatile}
\title{Versatile Remove Objects}
\usage{
rm.versatile(save.objects = NULL, save.patterns = NULL,
  rm.objects = NULL, rm.patterns = NULL, modes = list("integer",
  "double", "character", "list"), envir = .GlobalEnv, inherits = FALSE)
}
\arguments{
\item{save.objects}{an optional list naming objects to be saved (\emph{i.e.,}
not to be removed).}

\item{save.patterns}{an optional list of string patterns or
\link[base]{regex} which specifies the objects to be
saved (\emph{i.e.,} not to be removed).}

\item{rm.objects}{a list naming objects to be removed.}

\item{rm.patterns}{an optional list of string patterns or
\link[base]{regex} which specifies the objects to be
removed.}

\item{modes}{a list to set the type or storage mode of the save.patterns
and/or rm.patterns. For example "logical", "integer", "double",
"complex", "raw", "character", "list", "expression", "name",
"symbol" and "function".}

\item{envir}{the \link[base]{environment} to use. Default is set to
.GlobalEnv}

\item{inherits}{should the enclosing frames of the environment be inspected?}
}
\description{
Versatile function to remove objects.
}
\details{
The \link[base]{rm} function is not a versatile function because it
         cannot handle exceptional objects or string patterns to skip or
         remove. Moreover, \emph{rm} gives no warning before removing the
         objects and is not interactive. Also, the \code{pattern} argument
         of ls() object cannot handle multiple patterns or multiple
         \link[base:regex]{regular expressions}. This function has been
         developed for sake of versatility, where you can determine
         exceptional objects and patterns for saving and/or removing objects.
         Also, you receive proper message regarding the final approval of
         removing objects. The search strategy of rm.versatile is based upon
         \link[utils]{apropos} function; which enables you to determine the
         type or storage mode of the objects (\emph{e.g.,} double, integer,
         character, \emph{etc.}).
}
\examples{
# First let us making some objects
a <- 1
b <- "female"
c <- FALSE
d <- 4:10
e <- 5
f <- 6
g <- 7
df1 <- 8
df2 <- 9
dfcol <- 10
dff1 <- 11
dfff3 <- 12
mydata1 <- 13
mymydata <- 14
dataframe1 <- data.frame(
    gender = c("F", "M", "M", "F"),
    score = c(10, 12, 14, 18)
    )
matrix.a <- matrix(1:10, nrow = 2, ncol = 5)
matrix2 <- matrix(1:100, ncol = 2)
x <- c("a", "b", "c", "d")
y <- c("d", "e", "f")
# Remove objects by rm.versatile
rm.versatile(
    save.objects = list("a", "b", "c"),
    save.patterns = list("df", "data"),
    rm.objects = list("x", "y"),
    rm.patterns = list("matrix")
)
# List objects
ls()

}
